Add separate settings for each chat
